Removed SampleScene. Moved PlayerInput dot cs files to Scripts folder. -DF
