--- Assets/PlayerController.cs
@@ -1,36 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class PlayerController : MonoBehaviour
-{
-    private PlayerInputActions playerInput;
-
-    private Rigidbody2D rb;
-
-    [SerializeField] private float speed = 10f;
-    // Start is called before the first frame update
-    void Awake()
-    {
-        playerInput = new PlayerInputActions();
-        rb = GetComponent<Rigidbody2D>();
-    }
-
-    private void OnEnable()
-    {
-        playerInput.Enable();
-    }
-
-    private void OnDisable()
-    {
-        playerInput.Disable();
-    }
-
-    // Update is called once per frame
-    void FixedUpdate()
-    {
-        Vector2 moveInput = playerInput.Movement.Move.ReadValue<Vector2>();
-        rb.velocity = moveInput * speed;
-    }
-}
-
