+++ Assets/Scripts/PlayerInput.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 33197b178071b7f4fb8110f153af96d8
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
